% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeriveStatsForCube.R
\name{DeriveStatsForCube}
\alias{DeriveStatsForCube}
\title{Derive the descriptive statistics defined in a cube}
\usage{
DeriveStatsForCube(store, forsparqlprefix, domainName, dsdName, dataSet,
  deriveMeasureList = NULL, checkOnly = TRUE, validation.measure = NULL,
  myprefixes)
}
\arguments{
\item{store}{a RRDF store containing one or more date cubes}

\item{forsparqlprefix}{prefixes for SPARQL queries}

\item{domainName}{domainName for the RRDF cube}

\item{dsdName}{Dataset Descriptor Name}

\item{dataSet}{data.frame containing dataset to validate against/derive measures}

\item{deriveMeasureList}{TODO(mja): add code for this variabel}

\item{checkOnly}{TRUE if only the check is performed (default),
FALSE then measure is overwritten with the results}

\item{validation.measure}{list with property, prefix and
URIprefix. The derived results is stored in the cube in the store
using the property. This is only used is checkOnly==TRUE}

\item{myprefixes}{Prefixes used for storing the results - TODO(mja): combined with forsparqlprefix}
}
\value{
list with status TRUE if the operation was sucessfull
}
\description{
Derive the descriptive statistics defined in a cube
}

